+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+
+- Integration with vscode Test Explorer [#54][pr-54]
+
+### Changed
+
+- CellPath now uses subscript format `path/to/notebook.ipynb[Celln]` [#54][pr-54]
+
+[pr-54]: https://github.com/MusicalNinjaDad/pytest-ipynb2/pull/54
+
 ## [0.4.0] - 2025-03-03
 
 ### Fixed